test: #5170 rebuild sitemap tests for new structure
